pyproject.toml
[project]
name = "llm-log-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "pyyaml>=6.0.0",
]
